Fix TypeScript errors in transaction history table utils
- Update HistoryConnection type to use exact query result type
- Fix type compatibility issues with DepositAccountHistoryEntryConnection
- Regenerate GraphQL types to match current schema
