+++ apps/admin-panel/app/deposit-accounts/[deposit-account-id]/details-card/freeze-deposit-account/hooks/gql/deposit-account-freeze-mutation.ts
@@ -1,0 +1,20 @@
+import { gql } from "@apollo/client"
+
+export {
+    useDepositAccountFreezeMutation,
+    DepositAccountFreezeDocument,
+    GetDepositAccountDetailsPageDocument,
+    type DepositAccountFreezeMutation,
+    type DepositAccountFreezeMutationVariables,
+} from "@/lib/graphql/generated"
+
+gql`
+  mutation DepositAccountFreeze($input: DepositAccountFreezeInput!) {
+    depositAccountFreeze(input: $input) {
+      account {
+        id
+        depositAccountId
+      }
+    }
+  }
+`
