+++ apps/admin-panel/app/deposit-accounts/[deposit-account-id]/details-card/freeze-deposit-account/hooks/use-freeze-deposit-account.ts
@@ -1,23 +1,7 @@
 "use client"
 
-import { gql } from "@apollo/client"
+import { useDepositAccountFreezeMutation, GetDepositAccountDetailsPageDocument } from "./gql/deposit-account-freeze-mutation"
 
-import {
-  useDepositAccountFreezeMutation,
-  GetDepositAccountDetailsPageDocument,
-} from "@/lib/graphql/generated"
-
-
-gql`
-  mutation DepositAccountFreeze($input: DepositAccountFreezeInput!) {
-    depositAccountFreeze(input: $input) {
-      account {
-        id
-        depositAccountId
-      }
-    }
-  }
-`
 export function useFreezeDepositAccount() {
   const [freezeDepositAccount, { loading, reset }] = useDepositAccountFreezeMutation({
     refetchQueries: [GetDepositAccountDetailsPageDocument],
