+++ apps/admin-panel/app/deposit-accounts/[deposit-account-id]/details-card/unfreeze-deposit-account/hooks/gql/deposit-account-unfreeze-mutation.ts
@@ -1,0 +1,20 @@
+import { gql } from "@apollo/client"
+
+export {
+    useDepositAccountUnfreezeMutation,
+    DepositAccountUnfreezeDocument,
+    GetDepositAccountDetailsPageDocument,
+    type DepositAccountUnfreezeMutation,
+    type DepositAccountUnfreezeMutationVariables,
+} from "@/lib/graphql/generated"
+
+gql`
+  mutation DepositAccountUnfreeze($input: DepositAccountUnfreezeInput!) {
+    depositAccountUnfreeze(input: $input) {
+      account {
+        id
+        depositAccountId
+      }
+    }
+  }
+`
