+++ apps/admin-panel/app/deposit-accounts/[deposit-account-id]/hooks/gql/deposit-account-details-page-fragment.ts
@@ -28,4 +28,3 @@ export const DEPOSIT_ACCOUNT_DETAILS_PAGE_FRAGMENT = gql`
   }
   ${DEPOSIT_ACCOUNT_BALANCE_FRAGMENT}
 `
-
