+++ apps/admin-panel/app/deposit-accounts/[deposit-account-id]/transaction-history-table/hooks/gql/deposit-account-history-entry-fragment.ts
