+++ apps/admin-panel/app/deposit-accounts/[deposit-account-id]/transaction-history-table/utils.ts
@@ -1,7 +1,7 @@
 import { PaginatedData } from "@/components/paginated-table"
 import {
   DepositAccountHistoryEntryFieldsFragment,
-  DepositAccountHistoryConnectionFieldsFragment,
+  GetDepositAccountTransactionHistoryQuery,
 } from "@/lib/graphql/generated"
 
 export type HistoryNode = DepositAccountHistoryEntryFieldsFragment
@@ -14,14 +14,19 @@ export const VALID_ENTRY_TYPES = [
   "CancelledWithdrawalEntry",
 ] as const
 
-type HistoryConnection = DepositAccountHistoryConnectionFieldsFragment
+type DepositAccountFromQuery = Extract<
+  NonNullable<GetDepositAccountTransactionHistoryQuery["publicIdTarget"]>,
+  { __typename: "DepositAccount" }
+>
+
+type HistoryConnection = DepositAccountFromQuery["history"]
 
 export const filterValidEntries = (
   history: HistoryConnection | null,
 ): HistoryConnection["edges"] => history?.edges.filter(({ node }) => {
-    if (!node.__typename) return false
-    return VALID_ENTRY_TYPES.some((validType) => validType === node.__typename)
-  }) ?? []
+  if (!node.__typename) return false
+  return VALID_ENTRY_TYPES.some((validType) => validType === node.__typename)
+}) ?? []
 
 export const transformHistoryToPaginatedData = (
   history: HistoryConnection | null,
