--- apps/admin-panel/app/deposit-accounts/list/hooks/gql/fragments.ts
@@ -1,20 +1,0 @@
-import { gql } from "@apollo/client"
-
-import { DEPOSIT_ACCOUNT_BALANCE_FRAGMENT } from "@/lib/graphql/fragments/balance"
-
-export const DEPOSIT_ACCOUNT_LIST_FRAGMENT = gql`
-  fragment DepositAccountListFields on DepositAccount {
-    id
-    status
-    publicId
-    customer {
-      email
-      customerId
-    }
-    depositAccountId
-    balance {
-      ...DepositAccountBalanceFields
-    }
-  }
-  ${DEPOSIT_ACCOUNT_BALANCE_FRAGMENT}
-`
