+++ apps/admin-panel/lib/graphql/fragments/index.ts
@@ -1,0 +1,2 @@
+export { PAGE_INFO_FRAGMENT } from "./page-info"
+export { DEPOSIT_ACCOUNT_BALANCE_FRAGMENT } from "./balance"
